Updated folder structure for compute modals.
