+++ src/pages/workspaces/workspace/analysis/modals/CloudEnvironmentModal.ts
@@ -30,8 +30,8 @@ import {
   getCloudProviderFromWorkspace,
   GoogleWorkspace,
 } from 'src/libs/workspace-utils';
-import { AzureComputeModalBase } from 'src/pages/workspaces/workspace/analysis/modals/ComputeModal/AzureComputeModal';
-import { GcpComputeModalBase } from 'src/pages/workspaces/workspace/analysis/modals/ComputeModal/GcpComputeModal';
+import { AzureComputeModalBase } from 'src/pages/workspaces/workspace/analysis/modals/ComputeModal/AzureComputeModal/AzureComputeModal';
+import { GcpComputeModalBase } from 'src/pages/workspaces/workspace/analysis/modals/ComputeModal/GcpComputeModal/GcpComputeModal';
 import { CromwellModalBase } from 'src/pages/workspaces/workspace/analysis/modals/CromwellModal';
 import { GalaxyModalBase } from 'src/pages/workspaces/workspace/analysis/modals/GalaxyModal';
 import { HailBatchModal } from 'src/pages/workspaces/workspace/analysis/modals/HailBatchModal';
